Update README.md
Controller Area Network (CAN) is a communication system often used in vehicles to manage and prioritize messages sent by different sensors and components within the car. The goal is to demonstrate how CAN Message Priority works in such a system by simulating a car with sensors like an engine heat monitor and an obstacle finder, and a main control unit. The key focus is on how messages are ranked based on their importance, ensuring that the system handles them correctly and efficiently.